Feature/80 auth routes

+++ src/controllers/studentController.js
@@ -9,29 +9,37 @@ const Student = require('../models/Student');
 module.exports = {
   async updateStatus(request, response) {
     const { id } = request.params;
-    const { status } = request.body;
+    const { status, agentRole} = request.body;
 
     try {
-      const student = await Student.update({
-        status
-      }, {
-        where: {
-          id
+      if (agentRole === 1) {
+        const student = await Student.update({
+          status
+        }, {
+          where: {
+            id
+          }
+        });
+        if (student[0] === 0) {
+          return response.status(400).json(
+            {
+              message: 'Professor não encontrado!',
+            }
+          );
         }
-      });
-      if (student[0] === 0) {
-        return response.status(400).json(
+        return response.status(200).json(
           {
-            message: 'Professor não encontrado!',
+            data: student[0],
+            message: 'Atualizado com sucesso',
           }
         );
       }
-      return response.status(200).json(
-        {
-          data: student[0],
-          message: 'Atualizado com sucesso',
-        }
-      );
+      else{
+        return response.status(401).json({
+          message: 'O usuário não possui permissão para a ação'
+        }); 
+      }
+
     } catch (error) {
       console.log(error);
       return response.status(400).json(
@@ -45,25 +53,25 @@ module.exports = {
   async status(request, response) {
     const { status } = request.params;
     try {
-      const student = await Student.findAll({
-        where: { status }
-      });
+        const student = await Student.findAll({
+          where: { status }
+        });
 
-      if (student.length === 0) {
+        if (student.length === 0) {
+          return response.status(200).json(
+            {
+              message: 'Nenhum professor pendente',
+            }
+          );
+        }
         return response.status(200).json(
           {
-            message: 'Nenhum professor pendente',
+            data: {
+              student,
+            },
+            message: 'Professor encontrado',
           }
         );
-      }
-      return response.status(200).json(
-        {
-          data: {
-            student,
-          },
-          message: 'Professor encontrado',
-        }
-      );
     } catch (error) {
       console.log(error);
       return response.status(400).json(
@@ -180,6 +188,7 @@ module.exports = {
   async update(request, response) {
     const { id } = request.params;
     const {
+      agentRole,
       cpf,
       birthdate,
       institution,
@@ -193,31 +202,37 @@ module.exports = {
     } = request.body;
 
     try {
-      const student = await Student.update({
-        cpf,
-        birthdate,
-        institution,
-        grade,
-        cep,
-        number,
-        details,
-        description,
-        special,
-        status
-      }, {
-        where: {
-          id
-        }
-      });
+      if (agentRole === 1 || agentRole === 3) {
+        const student = await Student.update({
+          cpf,
+          birthdate,
+          institution,
+          grade,
+          cep,
+          number,
+          details,
+          description,
+          special,
+          status
+        }, {
+          where: {
+            id
+          }
+        });
 
-      if (student[0] === 0) {
-        return response.status(400).json({
-          message: 'Estudante não encontrado!'
+        if (student[0] === 0) {
+          return response.status(400).json({
+            message: 'Estudante não encontrado!'
+          });
+        }
+        return response.status(200).json({
+          data: student[0],
+          message: 'Atualizado com sucesso'
         });
       }
-      return response.status(200).json({
-        data: student[0],
-        message: 'Atualizado com sucesso'
+
+      return response.status(401).json({
+        message: 'O usuário não possui permissão para a ação'
       });
     } catch (error) {
       return response.status(404).json({
@@ -228,22 +243,28 @@ module.exports = {
 
   async delete(request, response) {
     const { id } = request.params;
-
+    const { agentRole } = request.body;
     try {
-      const student = await Student.destroy({
-        where: {
-          id
-        },
-      });
+      if (agentRole === 1 || agentRole === 3) {
+        const student = await Student.destroy({
+          where: {
+            id
+          },
+        });
 
-      if (student === 0) {
+        if (student === 0) {
+          return response.status(200).json({
+            message: 'Estudante não encontrado!'
+          });
+        }
         return response.status(200).json({
-          message: 'Estudante não encontrado!'
+          data: student,
+          message: 'Apagado com sucesso'
         });
       }
-      return response.status(200).json({
-        data: student,
-        message: 'Apagado com sucesso'
+
+      return response.status(401).json({
+        message: 'O usuário não possui permissão para a ação'
       });
     } catch (error) {
       console.log(error);
